# internal/web/subscriptions/api.gen.go
// Package subscriptions provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package subscriptions

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error string `json:"error"`
}

// Paging defines model for Paging.
type Paging struct {
	// Limit Limit items
	Limit *int `json:"limit,omitempty"`

	// Offset Offset number
	Offset *int `json:"offset,omitempty"`

	// Total Count of subscriptions
	Total *int `json:"total,omitempty"`
}

// Subscription defines model for Subscription.
type Subscription struct {
	// EndDate Дата окончания подписки (опционально)
	EndDate *string `json:"end_date"`

	// Id ID подписки
	Id openapi_types.UUID `json:"id"`

	// Price Стоимость месячной подписки в рублях
	Price int `json:"price"`

	// ServiceName Название сервиса, предоставляющего подписку
	ServiceName string `json:"service_name"`

	// StartDate Дата начала подписки (месяц и год)
	StartDate string `json:"start_date"`

	// UserId ID пользователя
	UserId openapi_types.UUID `json:"user_id"`
}

// SubscriptionRequest defines model for SubscriptionRequest.
type SubscriptionRequest struct {
	// EndDate Дата окончания подписки (опционально)
	EndDate *string `json:"end_date"`

	// Price Стоимость месячной подписки в рублях
	Price int `json:"price"`

	// ServiceName Название сервиса, предоставляющего подписку
	ServiceName string `json:"service_name"`

	// StartDate Дата начала подписки (месяц и год)
	StartDate string `json:"start_date"`

	// UserId ID пользователя
	UserId openapi_types.UUID `json:"user_id"`
}

// ListParams defines parameters for List.
type ListParams struct {
	// Limit Limit subscription items
	Limit int `form:"limit" json:"limit"`

	// Offset Offset subscription items
	Offset int `form:"offset" json:"offset"`
}

// SumParams defines parameters for Sum.
type SumParams struct {
	// UserId User ID
	UserId *openapi_types.UUID `form:"user_id,omitempty" json:"user_id,omitempty"`

	// ServiceName Service name
	ServiceName *string `form:"service_name,omitempty" json:"service_name,omitempty"`

	// Start Start date
	Start *string `form:"start,omitempty" json:"start,omitempty"`

	// End End date
	End *string `form:"end,omitempty" json:"end,omitempty"`

	// Limit Limit subscriptions for count price
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`

	// Offset Offset subscriptions
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// CreateJSONRequestBody defines body for Create for application/json ContentType.
type CreateJSONRequestBody = SubscriptionRequest

// UpdateJSONRequestBody defines body for Update for application/json ContentType.
type UpdateJSONRequestBody = SubscriptionRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List subscriptions
	// (GET /subscriptions)
	List(ctx echo.Context, params ListParams) error
	// Create subscription
	// (POST /subscriptions)
	Create(ctx echo.Context) error
	// List subscriptions with sum prices and filters
	// (GET /subscriptions/sum)
	Sum(ctx echo.Context, params SumParams) error
	// DeleteSubscription By ID
	// (DELETE /subscriptions/{id})
	Delete(ctx echo.Context, id openapi_types.UUID) error
	// Get subscription by id
	// (GET /subscriptions/{id})
	Get(ctx echo.Context, id string) error
	// Update subscription by id
	// (PUT /subscriptions/{id})
	Update(ctx echo.Context, id openapi_types.UUID) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// List converts echo context to params.
func (w *ServerInterfaceWrapper) List(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params ListParams
	// ------------- Required query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, true, "limit", ctx.QueryParams(), &params.Limit)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter limit: %s", err))
	}

	// ------------- Required query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, true, "offset", ctx.QueryParams(), &params.Offset)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter offset: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.List(ctx, params)
	return err
}

// Create converts echo context to params.
func (w *ServerInterfaceWrapper) Create(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Create(ctx)
	return err
}

// Sum converts echo context to params.
func (w *ServerInterfaceWrapper) Sum(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params SumParams
	// ------------- Optional query parameter "user_id" -------------

	err = runtime.BindQueryParameter("form", true, false, "user_id", ctx.QueryParams(), &params.UserId)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter user_id: %s", err))
	}

	// ------------- Optional query parameter "service_name" -------------

	err = runtime.BindQueryParameter("form", true, false, "service_name", ctx.QueryParams(), &params.ServiceName)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter service_name: %s", err))
	}

	// ------------- Optional query parameter "start" -------------

	err = runtime.BindQueryParameter("form", true, false, "start", ctx.QueryParams(), &params.Start)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter start: %s", err))
	}

	// ------------- Optional query parameter "end" -------------

	err = runtime.BindQueryParameter("form", true, false, "end", ctx.QueryParams(), &params.End)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter end: %s", err))
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", ctx.QueryParams(), &params.Limit)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter limit: %s", err))
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, false, "offset", ctx.QueryParams(), &params.Offset)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter offset: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Sum(ctx, params)
	return err
}

// Delete converts echo context to params.
func (w *ServerInterfaceWrapper) Delete(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Delete(ctx, id)
	return err
}

// Get converts echo context to params.
func (w *ServerInterfaceWrapper) Get(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Get(ctx, id)
	return err
}

// Update converts echo context to params.
func (w *ServerInterfaceWrapper) Update(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.Update(ctx, id)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/subscriptions", wrapper.List)
	router.POST(baseURL+"/subscriptions", wrapper.Create)
	router.GET(baseURL+"/subscriptions/sum", wrapper.Sum)
	router.DELETE(baseURL+"/subscriptions/:id", wrapper.Delete)
	router.GET(baseURL+"/subscriptions/:id", wrapper.Get)
	router.PUT(baseURL+"/subscriptions/:id", wrapper.Update)

}

type ListRequestObject struct {
	Params ListParams
}

type ListResponseObject interface {
	VisitListResponse(w http.ResponseWriter) error
}

type List200JSONResponse struct {
	Paging *Paging         `json:"paging,omitempty"`
	Rows   *[]Subscription `json:"rows,omitempty"`
}

func (response List200JSONResponse) VisitListResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type List400JSONResponse ErrorResponse

func (response List400JSONResponse) VisitListResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type List404JSONResponse ErrorResponse

func (response List404JSONResponse) VisitListResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type List500JSONResponse ErrorResponse

func (response List500JSONResponse) VisitListResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type CreateRequestObject struct {
	Body *CreateJSONRequestBody
}

type CreateResponseObject interface {
	VisitCreateResponse(w http.ResponseWriter) error
}

type Create201JSONResponse struct {
	Id *openapi_types.UUID `json:"id,omitempty"`
}

func (response Create201JSONResponse) VisitCreateResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type Create400JSONResponse ErrorResponse

func (response Create400JSONResponse) VisitCreateResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type Create500JSONResponse ErrorResponse

func (response Create500JSONResponse) VisitCreateResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type SumRequestObject struct {
	Params SumParams
}

type SumResponseObject interface {
	VisitSumResponse(w http.ResponseWriter) error
}

type Sum200JSONResponse struct {
	Paging   Paging         `json:"paging"`
	Rows     []Subscription `json:"rows"`
	TotalSum int            `json:"total_sum"`
}

func (response Sum200JSONResponse) VisitSumResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type Sum400JSONResponse ErrorResponse

func (response Sum400JSONResponse) VisitSumResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type Sum404JSONResponse ErrorResponse

func (response Sum404JSONResponse) VisitSumResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type Sum500JSONResponse ErrorResponse

func (response Sum500JSONResponse) VisitSumResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type DeleteRequestObject struct {
	Id openapi_types.UUID `json:"id"`
}

type DeleteResponseObject interface {
	VisitDeleteResponse(w http.ResponseWriter) error
}

type Delete204Response struct {
}

func (response Delete204Response) VisitDeleteResponse(w http.ResponseWriter) error {
	w.WriteHeader(204)
	return nil
}

type Delete400JSONResponse ErrorResponse

func (response Delete400JSONResponse) VisitDeleteResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type Delete404JSONResponse ErrorResponse

func (response Delete404JSONResponse) VisitDeleteResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type Delete500JSONResponse ErrorResponse

func (response Delete500JSONResponse) VisitDeleteResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type GetRequestObject struct {
	Id string `json:"id"`
}

type GetResponseObject interface {
	VisitGetResponse(w http.ResponseWriter) error
}

type Get200JSONResponse Subscription

func (response Get200JSONResponse) VisitGetResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type Get400JSONResponse ErrorResponse

func (response Get400JSONResponse) VisitGetResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type Get404JSONResponse ErrorResponse

func (response Get404JSONResponse) VisitGetResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type Get500JSONResponse ErrorResponse

func (response Get500JSONResponse) VisitGetResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type UpdateRequestObject struct {
	Id   openapi_types.UUID `json:"id"`
	Body *UpdateJSONRequestBody
}

type UpdateResponseObject interface {
	VisitUpdateResponse(w http.ResponseWriter) error
}

type Update200JSONResponse struct {
	Id *openapi_types.UUID `json:"id,omitempty"`
}

func (response Update200JSONResponse) VisitUpdateResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type Update400JSONResponse ErrorResponse

func (response Update400JSONResponse) VisitUpdateResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type Update404JSONResponse ErrorResponse

func (response Update404JSONResponse) VisitUpdateResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type Update500JSONResponse ErrorResponse

func (response Update500JSONResponse) VisitUpdateResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// List subscriptions
	// (GET /subscriptions)
	List(ctx context.Context, request ListRequestObject) (ListResponseObject, error)
	// Create subscription
	// (POST /subscriptions)
	Create(ctx context.Context, request CreateRequestObject) (CreateResponseObject, error)
	// List subscriptions with sum prices and filters
	// (GET /subscriptions/sum)
	Sum(ctx context.Context, request SumRequestObject) (SumResponseObject, error)
	// DeleteSubscription By ID
	// (DELETE /subscriptions/{id})
	Delete(ctx context.Context, request DeleteRequestObject) (DeleteResponseObject, error)
	// Get subscription by id
	// (GET /subscriptions/{id})
	Get(ctx context.Context, request GetRequestObject) (GetResponseObject, error)
	// Update subscription by id
	// (PUT /subscriptions/{id})
	Update(ctx context.Context, request UpdateRequestObject) (UpdateResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// List operation middleware
func (sh *strictHandler) List(ctx echo.Context, params ListParams) error {
	var request ListRequestObject

	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.List(ctx.Request().Context(), request.(ListRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "List")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(ListResponseObject); ok {
		return validResponse.VisitListResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// Create operation middleware
func (sh *strictHandler) Create(ctx echo.Context) error {
	var request CreateRequestObject

	var body CreateJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.Create(ctx.Request().Context(), request.(CreateRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Create")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(CreateResponseObject); ok {
		return validResponse.VisitCreateResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// Sum operation middleware
func (sh *strictHandler) Sum(ctx echo.Context, params SumParams) error {
	var request SumRequestObject

	request.Params = params

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.Sum(ctx.Request().Context(), request.(SumRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Sum")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(SumResponseObject); ok {
		return validResponse.VisitSumResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// Delete operation middleware
func (sh *strictHandler) Delete(ctx echo.Context, id openapi_types.UUID) error {
	var request DeleteRequestObject

	request.Id = id

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.Delete(ctx.Request().Context(), request.(DeleteRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Delete")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(DeleteResponseObject); ok {
		return validResponse.VisitDeleteResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// Get operation middleware
func (sh *strictHandler) Get(ctx echo.Context, id string) error {
	var request GetRequestObject

	request.Id = id

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.Get(ctx.Request().Context(), request.(GetRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Get")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetResponseObject); ok {
		return validResponse.VisitGetResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// Update operation middleware
func (sh *strictHandler) Update(ctx echo.Context, id openapi_types.UUID) error {
	var request UpdateRequestObject

	request.Id = id

	var body UpdateJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.Update(ctx.Request().Context(), request.(UpdateRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Update")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(UpdateResponseObject); ok {
		return validResponse.VisitUpdateResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}